+++ src/option.ts
@@ -63,30 +63,33 @@ export const marpCoreOptionForCLI = async ({ uri }: TextDocument) => {
   const parentFolder = uri.scheme === 'file' && path.dirname(uri.fsPath)
   const baseFolder = workspaceFolder ? workspaceFolder.uri.fsPath : parentFolder
 
-  if (baseFolder) {
-    const themeFiles: WorkFile[] = ((await Promise.all(
-      themes.loadStyles(Uri.parse(`file:${baseFolder}`)).map(promise =>
-        promise
-          .then(async theme => {
-            if (theme.type === ThemeType.File) {
-              return { path: theme.path, cleanup: () => Promise.resolve() }
-            }
+  const themeFiles: WorkFile[] = (
+    await Promise.all(
+      themes
+        .loadStyles(baseFolder ? Uri.parse(`file:${baseFolder}`) : undefined)
+        .map(promise =>
+          promise.then(
+            async theme => {
+              if (theme.type === ThemeType.File) {
+                return { path: theme.path, cleanup: () => Promise.resolve() }
+              }
 
-            if (theme.type === ThemeType.Remote) {
-              const cssName = `.marp-vscode-cli-theme-${nanoid()}.css`
-              const tmp = path.join(tmpdir(), cssName)
+              if (theme.type === ThemeType.Remote) {
+                const cssName = `.marp-vscode-cli-theme-${nanoid()}.css`
+                const tmp = path.join(tmpdir(), cssName)
 
-              await promisify(writeFile)(tmp, theme.css)
-              return { path: tmp, cleanup: () => promisify(unlink)(tmp) }
-            }
-          })
-          .catch(e => console.error(e))
-      )
-    )) as any).filter(w => w)
+                await promisify(writeFile)(tmp, theme.css)
+                return { path: tmp, cleanup: () => promisify(unlink)(tmp) }
+              }
+            },
+            e => console.error(e)
+          )
+        )
+    )
+  ).filter((w): w is WorkFile => !!w)
 
-    baseOpts.themeSet = themeFiles.map(w => w.path)
-    baseOpts.vscode.themeFiles = themeFiles
-  }
+  baseOpts.themeSet = themeFiles.map(w => w.path)
+  baseOpts.vscode.themeFiles = themeFiles
 
   return baseOpts
 }